requirements.txt
# Cryptographic signing (Phase 0.2 — HMAC-SHA256 + AES-256-GCM key storage + JWT bridge)
cryptography>=42.0.0
jcs>=0.3.0          # RFC 8785 canonical JSON — required for all JCS signing inputs
PyJWT>=2.8.0        # JWT Bearer token validation (uses cryptography backend for RS256/ES256)

# Async Database Drivers
redis>=5.0.1
asyncpg>=0.29.0
motor>=3.3.2

# Data Validation
pydantic>=2.5.3

# Environment
python-dotenv>=1.0.0

# MCP SDK
mcp>=1.2.0
starlette>=0.36.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
# SMTP alerts (lazy-imported in trimcp.notifications — safe to omit in minimal dev installs)
aiosmtplib>=3.0.1,<4

# Phase 2: Graphify — entity/relation extraction
spacy>=3.7.0
# After install run: python -m spacy download en_core_web_sm

# Phase 3.2: AST Parsing — consolidated grammars via tree-sitter-language-pack (Section 9.2).
# On Python 3.10+, newer pack releases may be available; 0.9.x supports Python 3.9+.
tree-sitter>=0.20.4
tree-sitter-language-pack>=0.9.0

# Phase 6: Async Job Queue
rq>=1.16.1

# Phase 2: Bridge subscription renewal (§10.7)
APScheduler>=3.10.0

# Phase 5: Jina Embeddings (jinaai/jina-embeddings-v2-base-code, 768-dim)
sentence-transformers>=2.3.1
transformers>=4.36.2
torch>=2.1.2

# Phase 7: Quad-Stack Media Ingestion
minio>=7.2.0

# Phase 1.2: Sleep consolidation — clustering (HDBSCAN)
scikit-learn>=1.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Phase 2b: Document format extraction (Appendix J)
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=1.0.0
extract-msg>=0.50.0
pypdf>=4.0.0
pdfminer.six>=20231228
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
chardet>=5.2.0
markdown-it-py>=3.0.0
selectolax>=0.3.20
striprtf>=0.0.27
defusedxml>=0.7.1
nbformat>=5.10.0
lxml>=5.0.0
PyYAML>=6.0.1

# Phase 2b Part 2 (Appendix J.14–J.18): diagrams, Adobe, CAD, project, encryption helpers
vsdx>=0.5.16
psd-tools>=1.9.28
ezdxf>=1.3.0
olefile>=0.47
msoffcrypto-tool>=5.4.1
